Move `free_variables` to `term.rs` and add more tests
