+++ src/de_bruijn.rs
@@ -6,7 +6,7 @@ use crate::term::{
         Quotient, Sum, True, Type, Unifier, Variable,
     },
 };
-use std::{cmp::Ordering, collections::HashSet, rc::Rc};
+use std::{cmp::Ordering, rc::Rc};
 
 // Shifting refers to increasing the De Bruijn indices of free variables. A cutoff determines which
 // variables are considered free. This operation is used to lower a term into a nested scope while
@@ -359,67 +359,11 @@ pub fn open<'a>(
     }
 }
 
-// Compute the free variables of a term. A cutoff determines which variables are considered free.
-// This function includes free variables in type annotations.
-pub fn free_variables<'a>(term: &Term<'a>, cutoff: usize, variables: &mut HashSet<usize>) {
-    match &term.variant {
-        Unifier(subterm) => {
-            if let Some(subterm) = &*subterm.borrow() {
-                free_variables(subterm, cutoff, variables);
-            }
-        }
-        Type | Integer | IntegerLiteral(_) | Boolean | True | False => {}
-        Variable(_, index) => {
-            if *index >= cutoff {
-                variables.insert(*index - cutoff);
-            }
-        }
-        Lambda(_, domain, body) => {
-            free_variables(domain, cutoff, variables);
-            free_variables(body, cutoff + 1, variables);
-        }
-        Pi(_, domain, codomain) => {
-            free_variables(domain, cutoff, variables);
-            free_variables(codomain, cutoff + 1, variables);
-        }
-        Application(applicand, argument) => {
-            free_variables(applicand, cutoff, variables);
-            free_variables(argument, cutoff, variables);
-        }
-        Let(definitions, body) => {
-            for (_, annotation, definition) in definitions {
-                free_variables(annotation, cutoff + definitions.len(), variables);
-                free_variables(definition, cutoff + definitions.len(), variables);
-            }
-
-            free_variables(body, cutoff + definitions.len(), variables);
-        }
-        Negation(subterm) => free_variables(subterm, cutoff, variables),
-        Sum(term1, term2)
-        | Difference(term1, term2)
-        | Product(term1, term2)
-        | Quotient(term1, term2)
-        | LessThan(term1, term2)
-        | LessThanOrEqualTo(term1, term2)
-        | EqualTo(term1, term2)
-        | GreaterThan(term1, term2)
-        | GreaterThanOrEqualTo(term1, term2) => {
-            free_variables(term1, cutoff, variables);
-            free_variables(term2, cutoff, variables);
-        }
-        If(condition, then_branch, else_branch) => {
-            free_variables(condition, cutoff, variables);
-            free_variables(then_branch, cutoff, variables);
-            free_variables(else_branch, cutoff, variables);
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use crate::{
         assert_same,
-        de_bruijn::{free_variables, open, shift},
+        de_bruijn::{open, shift},
         term::{
             Term,
             Variant::{
@@ -431,7 +375,7 @@ mod tests {
         },
     };
     use num_bigint::ToBigInt;
-    use std::{cell::RefCell, collections::HashSet, rc::Rc};
+    use std::{cell::RefCell, rc::Rc};
 
     #[test]
     fn shift_unifier_none() {
@@ -2079,449 +2023,5 @@ mod tests {
                 ),
             },
         );
-    }
-
-    #[test]
-    fn free_variables_unifier_none() {
-        let mut variables = HashSet::new();
-
-        free_variables(
-            &Term {
-                source_range: None,
-                variant: Unifier(Rc::new(RefCell::new(None))),
-            },
-            10,
-            &mut variables,
-        );
-
-        assert!(variables.is_empty());
-    }
-
-    #[test]
-    fn free_variables_unifier_some() {
-        let mut variables = HashSet::new();
-
-        free_variables(
-            &Term {
-                source_range: None,
-                variant: Unifier(Rc::new(RefCell::new(Some(Term {
-                    source_range: None,
-                    variant: Variable("x", 15),
-                })))),
-            },
-            10,
-            &mut variables,
-        );
-
-        assert!(variables.contains(&5));
-    }
-
-    #[test]
-    fn free_variables_type() {
-        let mut variables = HashSet::new();
-
-        free_variables(
-            &Term {
-                source_range: None,
-                variant: Type,
-            },
-            10,
-            &mut variables,
-        );
-
-        assert!(variables.is_empty());
-    }
-
-    #[test]
-    fn free_variables_variable_free() {
-        let mut variables = HashSet::new();
-
-        free_variables(
-            &Term {
-                source_range: None,
-                variant: Variable("x", 15),
-            },
-            10,
-            &mut variables,
-        );
-
-        assert!(variables.contains(&5));
-    }
-
-    #[test]
-    fn free_variables_variable_bound() {
-        let mut variables = HashSet::new();
-
-        free_variables(
-            &Term {
-                source_range: None,
-                variant: Variable("x", 5),
-            },
-            10,
-            &mut variables,
-        );
-
-        assert!(variables.is_empty());
-    }
-
-    #[test]
-    fn free_variables_lambda() {
-        let mut variables = HashSet::new();
-
-        free_variables(
-            &Term {
-                source_range: None,
-                variant: Lambda(
-                    "a",
-                    Rc::new(Term {
-                        source_range: None,
-                        variant: Variable("b", 15),
-                    }),
-                    Rc::new(Term {
-                        source_range: None,
-                        variant: Variable("b", 15),
-                    }),
-                ),
-            },
-            10,
-            &mut variables,
-        );
-
-        assert!(variables.contains(&4));
-        assert!(variables.contains(&5));
-    }
-
-    #[test]
-    fn free_variables_pi() {
-        let mut variables = HashSet::new();
-
-        free_variables(
-            &Term {
-                source_range: None,
-                variant: Pi(
-                    "a",
-                    Rc::new(Term {
-                        source_range: None,
-                        variant: Variable("b", 15),
-                    }),
-                    Rc::new(Term {
-                        source_range: None,
-                        variant: Variable("b", 15),
-                    }),
-                ),
-            },
-            10,
-            &mut variables,
-        );
-
-        assert!(variables.contains(&4));
-        assert!(variables.contains(&5));
-    }
-
-    #[test]
-    fn free_variables_application() {
-        let mut variables = HashSet::new();
-
-        free_variables(
-            &Term {
-                source_range: None,
-                variant: Application(
-                    Rc::new(Term {
-                        source_range: None,
-                        variant: Variable("b", 15),
-                    }),
-                    Rc::new(Term {
-                        source_range: None,
-                        variant: Variable("b", 16),
-                    }),
-                ),
-            },
-            10,
-            &mut variables,
-        );
-
-        assert!(variables.contains(&5));
-        assert!(variables.contains(&6));
-    }
-
-    #[test]
-    fn free_variables_let() {
-        let mut variables = HashSet::new();
-
-        free_variables(
-            &Term {
-                source_range: None,
-                variant: Let(
-                    vec![
-                        (
-                            "x",
-                            Rc::new(Term {
-                                source_range: None,
-                                variant: Type,
-                            }),
-                            Rc::new(Term {
-                                source_range: None,
-                                variant: Variable("y", 15),
-                            }),
-                        ),
-                        (
-                            "y",
-                            Rc::new(Term {
-                                source_range: None,
-                                variant: Type,
-                            }),
-                            Rc::new(Term {
-                                source_range: None,
-                                variant: Variable("z", 16),
-                            }),
-                        ),
-                    ],
-                    Rc::new(Term {
-                        source_range: None,
-                        variant: Variable("w", 17),
-                    }),
-                ),
-            },
-            10,
-            &mut variables,
-        );
-
-        assert!(variables.contains(&3));
-        assert!(variables.contains(&4));
-        assert!(variables.contains(&5));
-    }
-
-    #[test]
-    fn free_variables_integer() {
-        let mut variables = HashSet::new();
-
-        free_variables(
-            &Term {
-                source_range: None,
-                variant: Integer,
-            },
-            10,
-            &mut variables,
-        );
-
-        assert!(variables.is_empty());
-    }
-
-    #[test]
-    fn free_variables_integer_literal() {
-        let mut variables = HashSet::new();
-
-        free_variables(
-            &Term {
-                source_range: None,
-                variant: IntegerLiteral(ToBigInt::to_bigint(&84).unwrap()),
-            },
-            10,
-            &mut variables,
-        );
-
-        assert!(variables.is_empty());
-    }
-
-    #[test]
-    fn free_variables_negation() {
-        let mut variables = HashSet::new();
-
-        free_variables(
-            &Term {
-                source_range: None,
-                variant: Negation(Rc::new(Term {
-                    source_range: None,
-                    variant: Variable("b", 15),
-                })),
-            },
-            10,
-            &mut variables,
-        );
-
-        assert!(variables.contains(&5));
-    }
-
-    #[test]
-    fn free_variables_sum() {
-        let mut variables = HashSet::new();
-
-        free_variables(
-            &Term {
-                source_range: None,
-                variant: Sum(
-                    Rc::new(Term {
-                        source_range: None,
-                        variant: Variable("b", 15),
-                    }),
-                    Rc::new(Term {
-                        source_range: None,
-                        variant: Variable("b", 16),
-                    }),
-                ),
-            },
-            10,
-            &mut variables,
-        );
-
-        assert!(variables.contains(&5));
-        assert!(variables.contains(&6));
-    }
-
-    #[test]
-    fn free_variables_difference() {
-        let mut variables = HashSet::new();
-
-        free_variables(
-            &Term {
-                source_range: None,
-                variant: Difference(
-                    Rc::new(Term {
-                        source_range: None,
-                        variant: Variable("b", 15),
-                    }),
-                    Rc::new(Term {
-                        source_range: None,
-                        variant: Variable("b", 16),
-                    }),
-                ),
-            },
-            10,
-            &mut variables,
-        );
-
-        assert!(variables.contains(&5));
-        assert!(variables.contains(&6));
-    }
-
-    #[test]
-    fn free_variables_product() {
-        let mut variables = HashSet::new();
-
-        free_variables(
-            &Term {
-                source_range: None,
-                variant: Product(
-                    Rc::new(Term {
-                        source_range: None,
-                        variant: Variable("b", 15),
-                    }),
-                    Rc::new(Term {
-                        source_range: None,
-                        variant: Variable("b", 16),
-                    }),
-                ),
-            },
-            10,
-            &mut variables,
-        );
-
-        assert!(variables.contains(&5));
-        assert!(variables.contains(&6));
-    }
-
-    #[test]
-    fn free_variables_quotient() {
-        let mut variables = HashSet::new();
-
-        free_variables(
-            &Term {
-                source_range: None,
-                variant: Quotient(
-                    Rc::new(Term {
-                        source_range: None,
-                        variant: Variable("b", 15),
-                    }),
-                    Rc::new(Term {
-                        source_range: None,
-                        variant: Variable("b", 16),
-                    }),
-                ),
-            },
-            10,
-            &mut variables,
-        );
-
-        assert!(variables.contains(&5));
-        assert!(variables.contains(&6));
-    }
-
-    #[test]
-    fn free_variables_boolean() {
-        let mut variables = HashSet::new();
-
-        free_variables(
-            &Term {
-                source_range: None,
-                variant: Boolean,
-            },
-            10,
-            &mut variables,
-        );
-
-        assert!(variables.is_empty());
-    }
-
-    #[test]
-    fn free_variables_true() {
-        let mut variables = HashSet::new();
-
-        free_variables(
-            &Term {
-                source_range: None,
-                variant: True,
-            },
-            10,
-            &mut variables,
-        );
-
-        assert!(variables.is_empty());
-    }
-
-    #[test]
-    fn free_variables_false() {
-        let mut variables = HashSet::new();
-
-        free_variables(
-            &Term {
-                source_range: None,
-                variant: False,
-            },
-            10,
-            &mut variables,
-        );
-
-        assert!(variables.is_empty());
-    }
-
-    #[test]
-    fn free_variables_if() {
-        let mut variables = HashSet::new();
-
-        free_variables(
-            &Term {
-                source_range: None,
-                variant: If(
-                    Rc::new(Term {
-                        source_range: None,
-                        variant: Variable("b", 15),
-                    }),
-                    Rc::new(Term {
-                        source_range: None,
-                        variant: Variable("b", 16),
-                    }),
-                    Rc::new(Term {
-                        source_range: None,
-                        variant: Variable("c", 17),
-                    }),
-                ),
-            },
-            10,
-            &mut variables,
-        );
-
-        assert!(variables.contains(&5));
-        assert!(variables.contains(&6));
-        assert!(variables.contains(&7));
     }
 }
